requirements.txt
pip==21.2.3
setuptools==57.4.0
pyinstaller==6.3.0
requests==2.29.0
XlsxWriter==3.1.9
Jinja2==3.1.2
colorama==0.4.6
pika==1.3.2
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
importlib-metadata==7.0.1
MarkupSafe==2.1.3
packaging==23.2
pefile==2023.2.7
pyinstaller-hooks-contrib==2023.12
pywin32-ctypes==0.2.2
urllib3==1.26.18
zipp==3.17.0
python-fast-grpc
grpcio
grpcio-tools